Macro wt_systems::pid_step_tests

macro_rules! pid_step_tests {
    {
            name: $name:ident,
            config: $config:expr,
            initial: $initial:expr,
            steps: [
                $({
                    inputs: ($current:expr, $delta_t:expr),
                    expect: ($expected_output:expr, $expected_retained_error:expr)$(,)?
                }),*$(,)?
            ],
            tolerances: {
                output: $output_tolerance:expr,
                retained_error: $retained_error_tolerance:expr$(,)?
            }$(,)?
        } => { ... };
}

Provides a way to easily set up a test of a PID controller with certain initial conditions, and verify the behavior at each intermediate step