Trait wt_systems::pid::Pid

pub trait Pid<In> {
    type Configuration: Configuration;
    fn step_with_components(
        &mut self, 
        error: In, 
        config: &Self::Configuration, 
        plant_value: In, 
        delta_t: Time
    ) -> PidComponents;

    fn step(
        &mut self, 
        error: In, 
        config: &Self::Configuration, 
        plant_value: In, 
        delta_t: Time
    ) -> Ratio { ... }
}

A PID Controller

Associated Types

type Configuration: Configuration

The configuration type required for this PID

Loading content...
Required methods

fn step_with_components(
&mut self,
error: In,
config: &Self::Configuration,
plant_value: In,
delta_t: Time
) -> PidComponents

Steps the PID controller forward in time with intermediate outputs

There may be times where it is useful (perhaps due to a change in state of the system) to switch out the PID configuration. This can allow for switching to PIDs with different behavior that better match different phases.

Loading content...
Provided methods

fn step(
&mut self,
error: In,
config: &Self::Configuration,
plant_value: In,
delta_t: Time
) -> Ratio

Steps the PID controller forward in time

There may be times where it is useful (perhaps due to a change in state of the system) to switch out the PID configuration. This can allow for switching to PIDs with different behavior that better match different phases.

Loading content...